fix: lib-config points to bash_profile, use gnome keyring
lib-config used to point to .profile for Linux, but this is only needed
before you start the GUI. So put everything into bash_profile instead
to be more parallel to the Mac.

Change for lib-keychain to use the base gnome key ring rather than keychain
but fixed the code so it should work with it.