fix: new layout of profiel and rc files
Fixes problems with configuration and conforms Mac and Ubuntu
The main change is moving all non-interactive paths to .profile
and then putting all nonexportables such as functions, aliases and
history into bashrc. Finally, this moves all the interactive things like
veracrypt to .bashrc with a $- =~ i check to silence this if it is run
interactively.