fix(ci): nearly fixed github actions for local act
Tweaking the github workflow to run locally. Removed the snap
installation for a direct install by pre-commit. Since snap uses
services and act as of July 2022 does not support it.

Also removed the brew installation as this also fails with brew not
found. Bug submitted.

Finally the pre-commit action fails with no git command found. Submitted
bugs for both.