fix: shellcheck throwing SC1090 and SC1091 errors
Previously, each source in bash needed a shellcheck directive to tell it
where the file is. With shellcheck v0.8 this is no long needed, it
strips the prefix and tries for the local directory. This works well for
./bin and other scripts that are the root.

We also use the -x directive which allows external files to be searched.
That is scripts not explicitly on the shellcheck command line.

There are two ways this works. If you use docker, then you can use the
use the source shellcheck docker containers and set the version number.
This is slower on your main machine, but reliable.

The second way to do this is to use the jumanjihouse which uses the
system shellcheck. However, you need to brew install shellcheck and make
sure it is v0.8.0. This is now done with the GitHub Action which does a
brew install of the v0.8. Note that in the current Ubuntu default, it
has shellcheck 0.7 so we need to update to it.