fix(ci): pre-commit force logic and no autoupdate
Changed the force logic so that it actually works. $(FORCE) cannot be
inside a [[  ]] set it must just be evaluated as a shell command since
there are no real true or false in bash, they are just built-ins that
return different error codes.

Also removed the autoupdate from the pre-commit