fix: git install and update
Updated git install and update to echo to the stdout the actual path
where the repo is cloned so the caller does not have to make any
assumptions. Most of the time this is  $WS_DIR/git but this makes it
flexible if it moves.

there is a parallel commit in ./bin which normally does a pushd after
the call to get into the repo or pipes to /dev/null if you do not need
this.